feat/refactor: moved tools to agent folder and made multi-agent graph
