+++ backend/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "langchain-openai>=0.3.28",
     "langchain-tavily>=0.2.11",
     "langgraph>=0.6.3",
+    "langgraph-swarm>=0.0.14",
     "langsmith>=0.4.11",
     "requests>=2.32.4",
 ]
