+++ backend/src/optcg/agents/tools/__init__.py
@@ -2,9 +2,11 @@
 
 from .rulebook_tool import create_rulebook_retriever_tool
 from .get_board_tool import get_board_tool, get_board_tool_http
+from .handoff_tool import transfer_to_board_analyst
 
 __all__ = [
     "create_rulebook_retriever_tool",
     "get_board_tool",
-    "get_board_tool_http"
+    "get_board_tool_http",
+    "transfer_to_board_analyst"
 ]
