+++ backend/src/optcg/agents/tools/get_board_tool.py
