+++ backend/src/optcg/agents/tools/handoff_tool.py
@@ -1,0 +1,8 @@
+"""Agent tool to hand off tasks to other agents."""
+
+from langgraph_swarm import create_handoff_tool
+
+transfer_to_board_analyst = create_handoff_tool(
+    agent_name="board_analyst", 
+    description="Transfer the task to the board state analysis agent."
+)
