+++ backend/src/optcg/agents/tools/rulebook_tool.py
